Move plugin related stuff to its own package
